requirements.txt
fastapi[standard]
pydantic_settings
python-dotenv 
psycopg 
psycopg-binary
sqlmodel
requests